+++ src/main/java/racingcar/domain/CarRepository.java
@@ -1,10 +1,8 @@
-package racingcar.repository;
+package racingcar.domain;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
-
-import racingcar.domain.Car;
 
 public class CarRepository {
