+++ src/test/java/racingcar/util/AttemptValidatorTest.java
@@ -4,7 +4,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import racingcar.validator.AttemptValidator;
+import racingcar.util.validator.AttemptValidator;
 
 public class AttemptValidatorTest {
