+++ src/test/java/racingcar/util/InputValidatorTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import racingcar.validator.InputValidator;
+import racingcar.util.validator.InputValidator;
 
 public class InputValidatorTest {
